+++ src/capsula.ts
@@ -29,7 +29,8 @@ import rawDocker1              from "./capsula-container-alpine.dockerfile?raw" 
 import rawDocker2              from "./capsula-container-debian.dockerfile?raw" with { type: "string" }
 import rawDocker3              from "./capsula-container-ubuntu.dockerfile?raw" with { type: "string" }
 import rawDocker4              from "./capsula-container-alma.dockerfile?raw"   with { type: "string" }
-import rawDocker5              from "./capsula-container-arch.dockerfile?raw"   with { type: "string" }
+import rawDocker5              from "./capsula-container-fedora.dockerfile?raw" with { type: "string" }
+import rawDocker6              from "./capsula-container-arch.dockerfile?raw"   with { type: "string" }
 import rawBash                 from "./capsula-container.bash?raw"              with { type: "string" }
 import rawDefaults             from "./capsula.yaml?raw"                        with { type: "string" }
 
@@ -117,8 +118,8 @@ const spool = new Spool()
             type:     "string",
             coerce:   coerceS,
             default:  "debian",
-            choices:  [ "alpine", "debian", "ubuntu", "alma", "arch" ] as const,
-            describe: "set Linux platform (\"alpine\", \"debian\", \"ubuntu\", \"alma\", or \"arch\")"
+            choices:  [ "alpine", "debian", "ubuntu", "alma", "fedora", "arch" ] as const,
+            describe: "set Linux platform to use for container image"
         })
         .option("docker", {
             alias:    "d",
@@ -254,7 +255,8 @@ const spool = new Spool()
                     args.platform === "debian" ? rawDocker2 :
                         args.platform === "ubuntu" ? rawDocker3 :
                             args.platform === "alma" ? rawDocker4 :
-                                args.platform === "arch" ? rawDocker5 : ""
+                                args.platform === "fedora" ? rawDocker5 :
+                                    args.platform === "arch" ? rawDocker6 : ""
             await fs.promises.writeFile(dockerfile, dockerfileText, { encoding: "utf8" })
 
             /*  create entrypoint script  */