+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 ChangeLog
 =========
 
-0.9.9 (2025-11-28)
+0.9.9 (2025-11-29)
 ------------------
 
 - IMPROVEMENT: add Ubuntu Linux platform support
 - IMPROVEMENT: add Alma Linux platform support
 - IMPROVEMENT: add Arch Linux platform support
+- IMPROVEMENT: add Fedora Linux platform support
 
 0.9.8 (2025-11-27)
 ------------------